refactor: rename `@tarko/cli` to `@tarko/agent-cli`
