+++ multimodal/tarko/agent-cli/bin/cli.js
