+++ multimodal/tarko/agent-cli/src/config/paths.ts
