+++ multimodal/tarko/agent-cli/src/core/cli.ts
