+++ multimodal/tarko/agent-cli/src/core/commands/index.ts
