+++ multimodal/tarko/agent-cli/src/core/commands/run.ts
