+++ multimodal/tarko/agent-cli/src/core/commands/start.ts
