+++ multimodal/tarko/agent-cli/src/core/stdin.ts
