+++ multimodal/tarko/agent-cli/src/utils/console-interceptor.ts
