+++ multimodal/tarko/agent-cli/src/utils/logo.ts
